+++ src/extension.ts
@@ -2,96 +2,26 @@
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode'
 import * as gitExtension from './types/git'
-import { isDescendant } from './utils'
-import { getRemoteUrlPointLine, getTemplateFromRemote } from './link'
+import { ensurePosixPath, isDescendant } from './utils'
+import { getTemplateFromRemote } from './link'
 import { relative } from 'path'
 
-const SETTING_NAME = 'gitline'
+import {
+  defaultRemoteSelector,
+  currentRemoteSelector,
+  setDefaultRemote,
+} from './remote-selector'
 
-async function setDefaultRemote(
-  repository: gitExtension.Repository,
-  skipForOnlyOneRemote: boolean = false
-): Promise<string> {
-  if (repository.state.remotes.length === 1 && skipForOnlyOneRemote) {
-    return repository.state.remotes[0].name
-  }
-  const remote = await vscode.window.showQuickPick(
-    repository.state.remotes.map((r) => r.name),
-    {
-      placeHolder: 'Select remote',
-    }
-  )
-  if (!remote) {
-    throw new Error('remote is empty')
-  }
-  await vscode.workspace
-    .getConfiguration(SETTING_NAME)
-    .update('defaultRemote', remote, false)
-  return remote
-}
-
-async function defaultRemoteSelector(repository: gitExtension.Repository) {
-  let defaultRemote = vscode.workspace.getConfiguration(SETTING_NAME)
-    .defaultRemote as string
-  if (defaultRemote === '') {
-    defaultRemote = await setDefaultRemote(repository, true)
-  }
-  const selectedRemote =
-    repository.state.remotes.find((r) => r.name === defaultRemote) ??
-    repository.state.remotes[0]
-  return selectedRemote
-}
-
-function currentRemoteSelector(repository: gitExtension.Repository) {
-  const currentBranch = repository.state.HEAD!
-  return (
-    repository.state.remotes.find(
-      (r) => r.name === currentBranch.upstream?.remote
-    ) ?? repository.state.remotes[0]
-  )
-}
-
-async function getSpecParentCommit(
-  commit: gitExtension.Commit,
-  repository: gitExtension.Repository,
-  n: number
-): Promise<string> {
-  if (n === 0) {
-    return commit.hash
-  }
-  if (commit.parents.length === 0) {
-    return commit.hash
-  }
-  return getSpecParentCommit(
-    await repository.getCommit(commit.parents[0]),
-    repository,
-    n - 1
-  )
-}
-
-function currentBranchHeadSelector(
-  repository: gitExtension.Repository
-): () => Promise<string> {
-  return async () => {
-    const head = repository.state.HEAD!
-    if (head.ahead === 0) {
-      return head.commit!
-    }
-
-    return await getSpecParentCommit(
-      await repository.getCommit(head.commit!),
-      repository,
-      head.ahead!
-    )
-  }
-}
+import { getIdentifierGeneratorFromSetting } from './identifier-generator'
 
 function copyLineLinkCommand(
   repository: gitExtension.Repository,
   remoteSelector: (
     repository: gitExtension.Repository
   ) => Promise<gitExtension.Remote> | gitExtension.Remote,
-  branchOrCommitSelector: () => Promise<string> | string
+  branchOrCommitSelector: (
+    repository: gitExtension.Repository
+  ) => Promise<string> | string
 ): () => Promise<void> {
   return async () => {
     const activatedEditor = vscode.window.activeTextEditor
@@ -102,9 +32,16 @@ function copyLineLinkCommand(
       const currentPath = activatedEditor.document.uri
       const remote = await remoteSelector(repository)
       const template = getTemplateFromRemote(remote)
+
+      if (!currentPath.fsPath.startsWith(repository.rootUri.fsPath)) {
+        throw new Error('File is not in the repository')
+      }
+
       const lineLink = template({
-        relativePath: relative(repository.rootUri.fsPath, currentPath.fsPath),
-        branchOrCommit: await branchOrCommitSelector(),
+        relativePath: ensurePosixPath(
+          relative(repository.rootUri.fsPath, currentPath.fsPath)
+        ),
+        branchOrCommit: await branchOrCommitSelector(repository),
         beginLine: startLine + 1,
         endLine: endLine === startLine ? undefined : endLine + 1,
       })
@@ -124,7 +61,7 @@ function registerCommand(
       copyLineLinkCommand(
         repository,
         defaultRemoteSelector,
-        currentBranchHeadSelector(repository)
+        getIdentifierGeneratorFromSetting
       )
     ),
     vscode.commands.registerCommand(
@@ -132,13 +69,15 @@ function registerCommand(
       copyLineLinkCommand(
         repository,
         currentRemoteSelector,
-        currentBranchHeadSelector(repository)
+        getIdentifierGeneratorFromSetting
       )
     ),
     vscode.commands.registerCommand('git-line.set-default-remote', async () => {
       await setDefaultRemote(repository)
     })
   )
+
+  vscode.commands.executeCommand('setContext', 'gitline.enabled', true)
 }
 
 // This method is called when your extension is activated
@@ -170,4 +109,6 @@ export function activate(context: vscode.ExtensionContext) {
 }
 
 // This method is called when your extension is deactivated
-export function deactivate() {}
+export function deactivate() {
+  vscode.commands.executeCommand('setContext', 'gitline.enabled', false)
+}